ProcedureDefine allows definition of a new function with the same name as an already existing one provided the arguments are different
TypeDefine needs the arguments to the type input and output
functions to be 0, so ProcedureDefine automatically converts
them to 0.  these functions are detected by checking if the
type of an argument is actually defined or is just a shell.
this is a temporary hack; the whole mechanism of defining
functions with bogus types should be replaced.  i intend to
create a relation of functions whose domain and range are not
valid postquel types.