added constisset param to MakeConst calls added routines fixupsets() and define_sets() to go through parsetree, find Const nodes, convert query strings in them to set definitions added variable in parser() to save parsetree, so recursive calls to parser don't overwrite the parsetree we want.
in setup_tlist, found out whether the attribute was a set and used the
info to set ISCOMPLEX correctly.  If the attr is a set, it isn't complex
(even though its members may be).


MakeFunc call in ParseFunc had 9 parameters, should only have had
7.  Took out the extras.

added extra param to HandleNestedDots and ParseFunc: the current resno,
so that calls to ExpandAll inside ParseFunc can provide this