added constisset param to MakeConst calls
in ExpandAll, check if the attr is a set.  If it is, then it's not complex.


took out "rtype" variable in make_var - wasn't needed.  Had been used to
look up the typeid, when the typeid was already available.


added Paul's change to SkipToFromList to not skip past any beginning
of command keyword, i.e. append, delete, replace, retrieve  so the
from gets bound to the correct command

also changed the logic in SkipForwardToFromList so that Ch was restored
for all cases except finding a FROM